feat/Detail-DetailApi추가
